requirements.txt
SwissArmyTransformer>=0.4.9
numpy==1.24.4
transformers>=4.36.2
xformers>=0.0.22
torchvision>=0.16.2
spacy>=3.6.0
pillow>=10.2.0
deepspeed>=0.13.1
seaborn>=0.13.2
loguru~=0.7.2
streamlit>=1.31.0
timm>=0.9.12
accelerate>=0.26.1
pydantic>=2.6.0
boto3==1.34.89

# # for openai demo
# openai>=1.16.0
# sse-starlette>=1.8.2
# fastapi>=0.110.1
# httpx>=0.27.0
# uvicorn>=0.29.0
# jsonlines>=4.0.0
